test(unit testing): Add unit testing cases for core application services

+++ src/main/java/com/mcalzada/repository/CharacterRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CharacterRepository extends JpaRepository<Character, Integer>
+public interface CharacterRepository extends JpaRepository<Character, Long>
 {
 
     Optional<Character> findFirstByName(String name);
